+++ cedar-policy-core/src/error_macros.rs
@@ -34,3 +34,25 @@ macro_rules! impl_diagnostic_from_source_loc_field {
         }
     };
 }
+
+/// Macro which implements the `.labels()` and `.source_code()` methods of
+/// `miette::Diagnostic` by using the parameter `$i` which must be an `Expr`
+/// (or `Box<Expr>`) type field.
+#[macro_export]
+macro_rules! impl_diagnostic_from_expr_field {
+    ( $i:ident ) => {
+        fn source_code(&self) -> Option<&dyn miette::SourceCode> {
+            self.$i
+                .source_loc()
+                .as_ref()
+                .map(|loc| &loc.src as &dyn miette::SourceCode)
+        }
+
+        fn labels(&self) -> Option<Box<dyn Iterator<Item = miette::LabeledSpan> + '_>> {
+            self.$i.source_loc().as_ref().map(|loc| {
+                Box::new(std::iter::once(miette::LabeledSpan::underline(loc.span)))
+                    as Box<dyn Iterator<Item = _>>
+            })
+        }
+    };
+}
